+++ AppForSkills.Application/SkillPosts/Commands/DeleteSkillPost/DeleteSkillPostCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AppForSkills.Application.Common.Interfaces;
+using AppForSkills.Application.Exceptions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -22,6 +23,12 @@ namespace AppForSkills.Application.SkillPosts.Commands.DeleteSkillPost
             var skillPost = await _context.SkillPosts.Where(s => s.StatusId == 1 && s.Id == request.SkillPostId)
                 .Include(a => a.Ratings).Include(q => q.Comments)
                 .FirstOrDefaultAsync(cancellationToken);
+
+            if(skillPost == null)
+            {
+                throw new WrongIDException("Skill Post with gaved id could not delete, because not exists in database. " +
+                    "Give another id.");
+            }
 
             _context.SkillPosts.Remove(skillPost);
